module4 homework assignments 4.3 - 4.5 New commit

+++ src/module4/assignment_4_4/BankSystem.java
@@ -7,7 +7,7 @@ import module4.assignment_4_3.User;
  */
 public interface BankSystem {
 
-    void withdrawOfUser(User user, int amount);
+    void withdrawOfUser(User user, double amount);
     void fundUser(User user, int amount);
     void transferMoney(User fromUser, User toUser, int amount);
     void paySalary(User user);
